♻️ Rename delete-icons name
