+++ client/src/components/modules/todoItem.jsx
@@ -1,6 +1,6 @@
 import { useDispatch } from "react-redux";
 import check_icon from "../../assets/icons/check.svg";
-import delete_icon from "../../assets/icons/delete.svg";
+import delete_icon from "../../assets/icons/delete-icon.svg";
 import { changeTodoAsync, RemoveTodoAsync } from "../../store/actions/todo";
 
 const TodoItem = ({ todo }) => {
@@ -17,7 +17,7 @@ const TodoItem = ({ todo }) => {
   return (
     <li className="todo-item">
       <button className="check-btn" onClick={handleCheck}>
-        {'true' === todo.todo_checked && (
+        {"true" === todo.todo_checked && (
           <img className="check-btn-img" src={check_icon} alt="check icon" />
         )}
       </button>
